add explicit pipe access test
